Add logic of autoindex
