+++ internal/common/common.go
@@ -74,3 +74,10 @@ const (
 	// SegmentIndexPath storage path const for segment index files.
 	SegmentIndexPath = `index_files`
 )
+
+const (
+	IndexParamsKey = "params"
+	IndexTypeKey   = "index_type"
+	MetricTypeKey  = "metric_type"
+	DimKey         = "dim"
+)
